refactor: ProductRegister 패키지 이동 및 관련 파일 경로 수정
